fix: NetworkDefine 코드 병합
